% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mat.R
\name{d_mat}
\alias{d_mat}
\title{Calculate pairwise distance matrix}
\usage{
d_mat(coords)
}
\arguments{
\item{coords}{Provide a two-column array, matrix, or data.frame with xy coordinates of locations. Alternatively, an object of class \code{sf}.}
}
\value{
A square distance matrix
}
\description{
Calculate pairwise distance matrix
}
\examples{
set.seed(123)

## Create random coordinates
loc <- data.frame(x = runif(10,0,10),
                  y = runif(10,0,10))

## Create distance matrix
dist_mat <- d_mat(coords = loc)

}
\author{
Bill Peterman \href{mailto:peterman.73@osu.edu}{peterman.73@osu.edu}
}
