% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landconn-package.R
\docType{package}
\name{landconn-package}
\alias{landconn}
\alias{landconn-package}
\title{landconn: Landscape Connectivity}
\description{
Tools to calculate and explore landscape connectivity, developed for teaching and
research. The package covers two complementary approaches: patch-based incidence
function connectivity, and resistance-surface connectivity through 'Circuitscape'.
}
\section{Main functions}{

\describe{
\item{\code{\link[=dist_mat]{dist_mat()}}}{Build a pairwise (Euclidean) distance matrix from site
coordinates. Returns a \code{land_dist} object.}
\item{\code{\link[=ifc]{ifc()}}}{Calculate incidence function connectivity from a distance matrix,
with four model formulations. Returns an \code{ifc} object with \code{print}, \code{summary},
and \code{plot} methods.}
\item{\code{\link[=lower]{lower()}}}{Extract the lower triangle of a distance matrix as a vector.}
\item{\code{\link[=run_cs]{run_cs()}}}{Run 'Circuitscape' from R (via 'JuliaConnectoR') to produce
current maps or resistance distances across a resistance surface.}
\item{\code{\link[=omni_cs]{omni_cs()}}}{Create an omnidirectional current flow map following the
approach of Koen et al. (2014).}
\item{\code{\link[=julia_packages]{julia_packages()}}}{Start Julia from R and load the libraries the
'Circuitscape' tools depend on.}
}
}

\section{Getting started}{

See \code{vignette("getting-started", package = "landconn")} for a worked walkthrough.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wpeterman/landconn}
  \item Report bugs at \url{https://github.com/wpeterman/landconn/issues}
}

}
\author{
\strong{Maintainer}: Bill Peterman \email{peterman.73@osu.edu}

Authors:
\itemize{
  \item Bill Peterman \email{peterman.73@osu.edu}
}

}
\keyword{internal}
