% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic-methods.R
\name{logLik.ifc_optim}
\alias{logLik.ifc_optim}
\alias{nobs.ifc_optim}
\title{Log-likelihood and parameter count for an \code{ifc_optim} fit}
\usage{
\method{logLik}{ifc_optim}(object, ...)

\method{nobs}{ifc_optim}(object, ...)
}
\arguments{
\item{object}{An \code{ifc_optim} object from \code{\link[=ifc_optim]{ifc_optim()}}.}

\item{...}{Ignored.}
}
\value{
A \code{logLik} object. Its \code{df} attribute is the number of estimated parameters
(GLM coefficients, any dispersion parameter, plus one for \code{alpha}) and its \code{nobs}
attribute is the number of sites.
}
\description{
Returns the maximized log-likelihood with its degrees of freedom set to the GLM's
parameters plus one for the optimized \code{alpha}. This is what makes \code{\link[=AIC]{AIC()}}, \code{\link[=BIC]{BIC()}},
and \code{\link[=AICc]{AICc()}} count the scale parameter.
}
\seealso{
\code{\link[=AIC.ifc_optim]{AIC.ifc_optim()}}, \code{\link[=AICc]{AICc()}}, \code{\link[=ifc_modsel]{ifc_modsel()}}
}
