% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower.R
\name{lower}
\alias{lower}
\title{Extract the lower half of a distance matrix}
\usage{
lower(dist_mat)
}
\arguments{
\item{dist_mat}{Square distance matrix}
}
\value{
Vector of values from lower half of matrix
}
\description{
Extract the lower half of a distance matrix
}
\examples{
set.seed(123)

## Create random coordinates
loc <- data.frame(x = runif(10,0,10),
                  y = runif(10,0,10))

## Create distance matrix
d_mat <- dist_mat(coords = loc)

d_vec <- lower(d_mat)

}
\author{
Bill Peterman \href{mailto:peterman.73@osu.edu}{peterman.73@osu.edu}
}
