% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cs.R
\name{run_cs}
\alias{run_cs}
\title{Run Julia version of CIRCUITSCAPE from R}
\usage{
run_cs(JULIA_HOME = NULL,
rast,
input_locs = NULL,
regions = NULL,
field = NULL,
CurrentMap = TRUE,
VoltageMaps = FALSE,
Neighbor_Connect = 8,
full_mat = FALSE,
output_dir = NULL,
output_name = NULL,
output = "raster",
parallel = FALSE,
cores = NULL,
cholmod = TRUE,
precision = FALSE,
is_resistance = TRUE,
focal_node_current_zero = FALSE,
max_map = FALSE,
remove_files = TRUE,
silent = TRUE)
}
\arguments{
\item{JULIA_HOME}{Path to the folder containing the Julia binary (See Details).}

\item{rast}{Accepts a RasterLayer object or SpatRaster.}

\item{input_locs}{Provide a \code{\link[sp]{SpatialPoints}} object, \code{sf} point object, or \code{sf} polygon object.}

\item{regions}{If providing a point file for \code{input_locs}, specify short-circuit regions that overlap the points. Provide as \code{sf} polygon object. If \code{regions} are provided, your location points will not be used directly, but a random point from within each region will be used. (Default = NULL)}

\item{field}{(Default = NULL). Can be used to designate the identity of points or polygons when converting them to raster objects.If providing a polygon (SpatVector or sf POLYGON) this must be specified as the field name to get patch ID values from.}

\item{CurrentMap}{Logical. If TRUE, the cumulative current resistance map will be generated during the CS run (Default = TRUE)}

\item{cumulative_map_only}{Only write individual pairwise current maps, or also write each pairwise current map? (Default = TRUE)}

\item{VoltageMaps}{Logical. If TRUE, voltage maps will be produced.}

\item{scenario}{Specify 'pairwise' (Default), 'one-to-all', 'all-to-one', or 'advanced'. See CIRCUITSCAPE guide for details on usage.}

\item{Neighbor_Connect}{Specify 4 or 8 neighbor connection scheme (Default = 8)}

\item{output_dir}{Directory where CS results will be written. Only specify if Circuitscape current map outputs are requested. It is critical that there are NO SPACES in the directory, as this may cause the function to fail.}

\item{output_name}{Name to be applied to output files. Defaults to using the raster layer names.}

\item{output}{Specify either "matrix" or "raster". "matrix" will return the pairwise resistance matrix, while "raster" (Default) will return a \code{spatRaster} or \code{RasterLayer} object of the cumulative current map. The raster map can only be returned if \code{CurrentMap=TRUE}}

\item{pairs_to_include}{Default is NULL. If you wish to use the advanced CIRCUITSCAPE setting mode to include or exclude certain pairs of sample locations, provide a vector consisting of 1 (keep) or 0 (drop) for each pairwise observation (see example). This is an option if you do not want to assess all pairwise observations.}

\item{variable_source}{When using 'one-to-all', 'all-to-one', or 'advanced' modes. Provide a two-column data frame or matrix with the first column being point/patch ID and the second column being the input source strength. Must be specified in conjunction with \code{variable_ground} in advanced mode. Can be specified by itself in 'one-to-all' and 'all-to-one'.}

\item{variable_ground}{Required when using 'advanced' mode. Provide a two-column data frame or matrix with the first column being point/patch ID and the second column being in the input source strength. Must be specified in conjunction with \code{variable_source} in advanced mode.}

\item{parallel}{(Logical; Default = FALSE) Do you want to run CIRCUITSCAPE in parallel? Only consider when dealing with rasters >2 million cells and if using Mac or Linux OS.}

\item{cores}{If \code{parallel = TRUE}, how many cores should be used for parallel processing?}

\item{cholmod}{(Logical; Default = TRUE). Should the cholmod solver be used?}

\item{precision}{(Logical; Default = FALSE). Should experimental single precision method be used? See details.}

\item{is_resistance}{Default = TRUE. Is the landscape represented as a resistance (TRUE) or conductance (FALSE) surface?}

\item{focal_node_current_zero}{Set focal node current to 0 (Default = FALSE). May help with visualization of current flows in mapping.}

\item{max_map}{Should maximum current map be created and saved (Default = FALSE)?}

\item{remove_files}{Remove temporary files. Default is TRUE.}

\item{silent}{Printing of output from Circuitscape will be suppressed unless an error occurs (Default = TRUE)}
}
\value{
Function will save all output files created by CIRCUITSCAPE. Can choose to have a Full square distance matrix or Raster objects of the current maps returned to R. Raster object will be SpatRaster or RasterLayer, depending upon input format.
}
\description{
Execute julia CS from R
}
\details{
There is extensive documentation for Circuitscape here: https://docs.circuitscape.org/Circuitscape.jl/latest/ . Not all features and functionality have been built into this R function.
}
\examples{
## To do

## Not run:
## *** TO BE COMPLETED *** ##

## End (Not run)
}
\author{
Bill Peterman \href{mailto:peterman.73@osu.edu}{peterman.73@osu.edu}
}
