% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_function.R
\name{ifc}
\alias{ifc}
\title{Calculate connectivity using incidence functions}
\usage{
ifc(alpha,
           dist_mat,
           model,
           occ_sites = NULL,
           patch_area = NULL,
           scale = FALSE)
}
\arguments{
\item{alpha}{The average dispersal distance of your organism. Should be in the same units as the distances in the \code{dist_mat}. See Details}

\item{dist_mat}{A square pairwise distance matrix}

\item{model}{Numeric value 1-4 corresponding to how connectivity will be calculated. See Details}

\item{occ_sites}{A 0/1 vector equal in length to the number of sites in the \code{dist_mat}. If provided, only sites indicated with a 1 will contribute to connectivity. If not provided, all sites contribute to connectivity.}

\item{patch_area}{If provided, can be used calculate connectivity as a function of surrounding patched (Model 2), focal patches (Model 3), or both (Model 4). See details}

\item{scale}{Logical (Default = FALSE). If TRUE, connectivity estimates will be scaled to have a maximum of 1.}
}
\value{
A vector of relative connectivity values for each site
}
\description{
Calculate connectivity using incidence functions
}
\details{
\code{alpha} is used in conjunction with \code{dist_mat} in a negative exponential equation to calculate pairwise probability of connectivity:
\deqn{C ~ exp(-(1 / alpha) * dist_mat)}.
\subsection{CONNECTIVITY MODELS}{

model 1 --> Distance only \cr
Calculate connectivity as a function of distance.\cr
Hypothesis: Connectivity of patches declines as they become more isolated from other patches
\deqn{C_i = \sum_{j \ne i }p_j \text{exp} (-\alpha d_{ij})}

model 2 --> Contributing area \cr
Calculate connectivity, weighting by the area of contributing patches.\cr
Hypothesis: Larger patches contribute more individuals
\deqn{C_i = \sum_{j \ne i }p_j \text{exp} (-\alpha d_{ij})A_j}

model 3 --> Focal area \cr
Calculate connectivity, weighting by the area of the focal patches.\cr
Hypothesis: Larger patches have greater attraction
\deqn{C_i = A_i\sum_{j \ne i }p_j \text{exp} (-\alpha d_{ij})}

model 4 --> Contributing & Focal area \cr
Hypothesis: Larger patches contribute more individuals and larger patches have greater attraction \cr
\deqn{C_i = A_i\sum_{j \ne i }p_j \text{exp} (-\alpha d_{ij})A_j}

\eqn{p_j} (\code{occ_sites}) is the occupancy of site j and modifies which patches contribute to the connectivity of a patch
}
}
\examples{
set.seed(123)

## Create random coordinates
loc <- data.frame(x = runif(10,0,10),
                  y = runif(10,0,10))

## Create distance matrix
dist_mat <- as.matrix(dist(loc))

## Set alpha and indicate occupied sites
alpha <- 3
occ_sites <- c(0,0,1,1,0,1,0,0,1,1)

## Patch Areas
p_a <- runif(10, 5, 50)


## Calculate connectivity
## Model 1
(c1 <- ifc(alpha,
           dist_mat,
           model = 1))

## Model 1, scaled
(c1s <- ifc(alpha,
            dist_mat,
            model = 1,
            scale = TRUE))

## Model 2
(c2 <- ifc(alpha,
           dist_mat,
           model = 2,
           patch_area = p_a))

## Model 3
(c3 <- ifc(alpha,
           dist_mat,
           model = 3,
           patch_area = p_a))

## Model 4
(c4 <- ifc(alpha,
           dist_mat,
           model = 4,
           patch_area = p_a))

## Model 4, Occupied sites only
(c4b <- ifc(alpha,
            dist_mat,
            model = 4,
            occ_sites = occ_sites,
            patch_area = p_a))

}
\author{
Bill Peterman \href{mailto:peterman.73@osu.edu}{peterman.73@osu.edu}
}
